[S12P21A701-998] refactor: Change data type of subscription period from Long to Integer

+++ backend/src/main/java/bank/donghang/core/accountproduct/dto/request/AccountProductCreationRequest.java
@@ -10,7 +10,7 @@ public record AccountProductCreationRequest(
 	Double interestRate,
 	String rateDescription,
 	Integer accountProductTypeCode,
-	Long subscriptionPeriod,
+	Integer subscriptionPeriod,
 	Long minSubscriptionBalance,
 	Long maxSubscriptionBalance
 ) {
